added chatbot logic, sessions, clients routes, client controller, implemented client chatting

+++ controllers/question.js
@@ -1,4 +1,3 @@
-import { ObjectId } from "bson";
 import { QuestionModel } from "../models/question";
 import { SettingsModel } from "../models/settings";
 
@@ -7,8 +6,15 @@ export class Question {
     const questions = await QuestionModel.find().lean();
     const settings = await SettingsModel.findOne().lean();
     if (settings?.questionsOrder?.length) {
-      const reorderedQuestions = settings.questionsOrder.map((id) =>
-        questions.find((question) => question._id === id)
+      const reorderedQuestions = Array.from(
+        new Set([
+          ...settings.questionsOrder
+            .map((question) =>
+              questions.find((q) => String(q._id) === question)
+            )
+            .filter((el) => el),
+          ...questions,
+        ])
       );
       return reorderedQuestions;
     }
@@ -42,7 +48,6 @@ export class Question {
         status: 404,
       };
     }
-    console.log({ question });
     block.title = title;
     block.color = color;
     block.question = question;
@@ -52,17 +57,25 @@ export class Question {
     await block.save();
     return this.get();
   }
-  static async reorder({ ids }) {
+  static async reorder({ questionsIds }) {
+    if (!questionsIds)
+      throw {
+        message: "IDs das perguntas não informados",
+      };
     await SettingsModel.findOneAndUpdate(
       {},
       {
-        $set: { questionsOrder: ids },
+        $set: { questionsOrder: questionsIds },
       },
       { upsert: true }
     );
     return await this.get();
   }
   static async delete({ id }) {
+    if (!id)
+      throw {
+        message: "ID não informado",
+      };
     await QuestionModel.findByIdAndDelete(id);
     return await this.get();
   }
